+++ supreme/common/src/jvmMain/kotlin/YamlFlatteningPolymorphicSerializer.kt
@@ -14,9 +14,7 @@ import kotlinx.serialization.json.buildJsonArray
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.doubleOrNull
 import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.longOrNull
-import kotlinx.serialization.json.put
 import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.PolymorphicSerializer
 import net.mamoe.yamlkt.YamlElement